requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
neo4j
faiss-cpu
python-dotenv
pandas
sentence-transformers
numpy
